[ADH-Funnel-872]: create E11 Front - feat standalone google autocomplete
