+++ src/Twig/Components/ReGoogleAutoComplete.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace App\Twig\Components;
+
+use App\Twig\AbstractComponentsLogic;
+use Symfony\UX\TwigComponent\Attribute\AsTwigComponent;
+
+#[AsTwigComponent]
+class ReGoogleAutoComplete extends AbstractComponentsLogic
+{
+    use StatusAwareTrait;
+
+    public ?string $label;
+    public ?string $id;
+    public array $searchBoxProps = [];
+    public string $associatedFieldsPrefix;
+}
