+++ templates/components/ReGoogleAutoComplete.html.twig
@@ -1,0 +1,18 @@
+
+{% block final_javascripts %}
+<script async defer>
+    const init_ac_{{ id }} = async () => await Main.initGooglePlacesAutocomplete({
+        apiKey: '{{ google_maps_api_key }}',
+        associatedFieldsPrefix: '{{ associatedFieldsPrefix }}',
+        autocompleteInputId: '{{ id }}',
+    })
+</script>
+{% endblock %}
+
+<div x-init="await init_ac_{{ id }}()">
+<twig:Molecules:ReSelect
+    {{ ...attributes.all() }}
+    onQuery="(x) => window.queryGooglePlaces(x)"
+    id="{{ id }}"
+    blocked="blocked" />
+</div>
