+++ templates/renaissance/adhesion/member_card.html.twig
@@ -1,0 +1,43 @@
+{% extends 'base_renaissance.html.twig' %}
+
+{% block page_title 'Création de mot de passe' %}
+
+{% block flash_message '' %}
+
+{#{% form_theme form 'renaissance/forms/tailwind_form_theme.html.twig' %}#}
+
+{% block content %}
+    <twig:ReStandaloneLayout class="re_new_ui">
+        <twig:Atoms:ReCard class="max-w-[500px]">
+{#            {% include 'renaissance/partials/flashes.html.twig' with {slim: true} %}#}
+
+            <form class="flex flex-col gap-8">
+                <div>
+                    <twig:Atoms:ReLabel> Carte d'adhérant</twig:Atoms:ReLabel>
+                    <p>
+                        Votre carte d’adhérent dématérialisée vient d’être générée rien que pour vous.
+                    </p>
+
+                </div>
+                <div>
+                    <twig:Atoms:ReLabel>Adresse Postale</twig:Atoms:ReLabel>
+                    <p>Nous utiliserons cette adresse pour vous envoyer votre reçu fiscal et votre carte d’adhérent.</p>
+                </div>
+
+                <twig:ReGoogleAutoComplete
+                    id="exemple_address_autocomplete"
+                    associatedFieldsPrefix="exemple_address"
+                    placeholder="Indiquez votre adresse"
+                />
+
+                <input type="text" hidden id="exemple_address_address">
+                <input type="text" hidden id="exemple_address_postalCode">
+                <input type="text" hidden id="exemple_address_cityName">
+                <input type="text" hidden id="exemple_address_country">
+
+                <twig:ReButton type="submit">Recevoir ma carte</twig:ReButton>
+                <twig:ReButton link>Je ne veux pas recevoir ma carte</twig:ReButton>
+            </form>
+        </twig:Atoms:ReCard>
+    </twig:ReStandaloneLayout>
+{% endblock %}
